I Fixed the localization provider .dart file and code for betther usability and folder place!
